+++ DependencyInjection/DarvinMenuExtension.php
@@ -73,6 +73,7 @@ class DarvinMenuExtension extends Extension implements PrependExtensionInterface
         (new ExtensionConfigurator($container, __DIR__.'/../Resources/config/app'))->configure([
             'darvin_admin',
             'darvin_utils',
+            'doctrine',
             'knp_menu',
             'twig',
         ]);